# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/candidates/update-personal-summary/candidate/{id}": {
      "put": {
        "tags": [
          "candidate-controller"
        ],
        "operationId": "updatePersonalSummary",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateResponse"
                }
              }
            }
          }
        }
      }
    },
    "/candidates/update-personal-infos/candidate/{id}": {
      "put": {
        "tags": [
          "candidate-controller"
        ],
        "operationId": "updatePersonalInfos",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateUpdateDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateResponse"
                }
              }
            }
          }
        }
      }
    },
    "/candidates/upadate/add-habilities/candidate/{id}": {
      "put": {
        "tags": [
          "candidate-controller"
        ],
        "operationId": "addHabilities",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateResponse"
                }
              }
            }
          }
        }
      }
    },
    "/vacancys/{empresaId}/create-vacancy": {
      "post": {
        "tags": [
          "vacancy-controller"
        ],
        "operationId": "createVacancy",
        "parameters": [
          {
            "name": "empresaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VacancyCreateDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Vacancy"
                }
              }
            }
          }
        }
      }
    },
    "/users/register/enterprise": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "operationId": "insertEnterprise",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreateEnterpriseDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/users/register/candidate": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "operationId": "insertCandidate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreateDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/skills/create": {
      "post": {
        "tags": [
          "skill-base-controller"
        ],
        "operationId": "insert",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillBase"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SkillBase"
                }
              }
            }
          }
        }
      }
    },
    "/selection-process/{vacancyId}/create-selection-process": {
      "post": {
        "tags": [
          "selection-process-controller"
        ],
        "operationId": "createVacancy_1",
        "parameters": [
          {
            "name": "vacancyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectionProcessCreateDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SelectionProcessResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/vacancys/{id}": {
      "get": {
        "tags": [
          "vacancy-controller"
        ],
        "operationId": "findById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Vacancy"
                }
              }
            }
          }
        }
      }
    },
...
    "/users/{id}": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "operationId": "findById_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
...
    "/skills/{id}": {
      "get": {
        "tags": [
          "skill-base-controller"
        ],
        "operationId": "findById_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SkillBase"
                }
              }
            }
          }
        }
      }
    },
...
    "/selection-process/{id}": {
      "get": {
        "tags": [
          "selection-process-controller"
        ],
        "operationId": "findById_3",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SelectionProcess"
                }
              }
            }
          }
        }
      }
    },
...
    "/candidates/{id}": {
      "get": {
        "tags": [
          "candidate-controller"
        ],
        "operationId": "findById_4",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateResponse"
                }
              }
            }
          }
        }
      }
    },
    "/candidates/findCandidateByUser/{userId}": {
      "get": {
        "tags": [
          "candidate-controller"
        ],
        "operationId": "findByUserId",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Enterprise": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "enterpriseName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "cnpj": {
            "type": "string"
          },
          "contact": {
            "type": "string"
          },
          "socialReason": {
            "type": "string"
          },
          "locality": {
            "type": "string"
          },
          "sizeEnterprise": {
            "type": "string",
            "enum": [
              "SMALL",
              "MEDIUM",
              "BIG"
            ]
          },
          "numberOfEmployees": {
            "type": "integer",
            "format": "int64"
          },
          "sector": {
            "type": "string"
          },
          "siteUrl": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "approved": {
            "type": "boolean"
          }
        }
      },
      "Role": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "roleName": {
            "type": "string",
            "enum": [
              "ADMINISTRADOR",
              "CANDIDATO",
              "EMPRESA"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Vacancy": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "enterprise": {
            "$ref": "#/components/schemas/Enterprise"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "position": {
            "type": "string"
          },
          "vacancyType": {
            "type": "string",
            "enum": [
              "FULL_TIME",
              "PART_TIME",
              "INTERSHIP",
              "TRAINEE",
              "TEMPORARY",
              "CONTRACT",
              "FREELANCE",
              "SELF_EMPLOYED",
              "VOLUNTEER",
              "APPRENTICESHIP"
            ]
          },
          "locality": {
            "type": "string"
          },
          "modalityType": {
            "type": "string",
            "enum": [
              "ON_SITE",
              "REMOTE",
              "HYBRID"
            ]
          },
          "publicationDate": {
            "type": "string",
            "format": "date"
          },
          "expirationDate": {
            "type": "string",
            "format": "date"
          },
          "closedAt": {
            "type": "string",
            "format": "date-time"
          },
          "vacancyStatus": {
            "type": "string",
            "enum": [
              "PENDING_APPROVAL",
              "APPROVED",
              "OPEN",
              "CLOSED",
              "PAUSED",
              "ARCHIVED",
              "REJECTED"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CandidateResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "fullName": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "raceEthnicity": {
            "type": "string"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date"
          },
          "cpf": {
            "type": "string"
          },
          "contact": {
            "type": "string"
          },
          "profileResponse": {
            "$ref": "#/components/schemas/ProfileResponse"
          }
        }
      },
      "ProfileResponse": {
        "type": "object",
        "properties": {
          "photo": {
            "type": "string"
          },
          "bannerm": {
            "type": "string"
          },
          "headline": {
            "type": "string"
          },
          "personalSummary": {
            "type": "string"
          },
          "socialLinks": {
            "type": "string"
          },
          "locality": {
            "type": "string"
          }
        }
      },
      "CandidateUpdateDTO": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "raceEthnicity": {
            "type": "string"
          },
          "headLine": {
            "type": "string"
          },
          "sector": {
            "type": "string"
          },
          "locality": {
            "type": "string"
          },
          "socialLinks": {
            "type": "string"
          },
          "contact": {
            "type": "string"
          }
        }
      },
      "VacancyCreateDTO": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "position": {
            "type": "string"
          },
          "vacancyType": {
            "type": "string"
          },
          "locality": {
            "type": "string"
          },
          "modalityType": {
            "type": "string"
          },
          "skillsIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
      "UserCreateEnterpriseDTO": {
        "type": "object",
        "properties": {
          "enterpriseName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "cnpj": {
            "type": "string"
          },
          "contact": {
            "type": "string"
          }
        }
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "email": {
            "type": "string"
          }
        }
      },
      "UserCreateDTO": {
        "type": "object",
        "properties": {
          "candidateName": {
            "type": "string"
          },
          "cpf": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "contact": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "SkillBase": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "skillType": {
            "type": "string",
            "enum": [
              "BEHAVIORAL",
              "TECHNICAL"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "SelectionProcessCreateDTO": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "SelectionProcessResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "vacancy": {
            "$ref": "#/components/schemas/VacancyResponse"
          }
        }
      },
      "VacancyResponse": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "position": {
            "type": "string"
          },
          "locality": {
            "type": "string"
          },
          "publicationDate": {
            "type": "string",
            "format": "date"
          },
          "modalityType": {
            "type": "string"
          },
          "vacancyStatus": {
            "type": "string"
          },
          "vacancyType": {
            "type": "string"
          },
          "expirationDate": {
            "type": "string",
            "format": "date"
          },
          "closedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "SelectionProcess": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "vacancy": {
            "$ref": "#/components/schemas/Vacancy"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "active": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}