# src/main/java/com/example/demo/dto/CandidateUpdateDTO.java
package com.example.demo.dto;

import lombok.Getter;

import java.time.LocalDate;

@Getter
public class CandidateUpdateDTO {
    private String fullName;
    private String gender;
    private String raceEthnicity;
    private String headLine;
    private String sector;
    private String locality;
    private String socialLinks;
    private String contact;
}