# src/main/java/com/example/demo/dto/UserCreateEnterpriseDTO.java
package com.example.demo.dto;

import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
public class UserCreateEnterpriseDTO {
    private String enterpriseName;
    private String email;
    private String cnpj;
    private String contact;
}