# src/main/java/com/example/demo/dto/VacancyCreateDTO.java
package com.example.demo.dto;

import com.example.demo.enums.VacancyType;
import com.example.demo.enums.WorkModality;
import lombok.Getter;

import java.util.List;

@Getter
public class VacancyCreateDTO {

    private String title;
    private String description;
    private String position;
    private String vacancyType;
    private String locality;
    private String modalityType;
    private List<Long> skillsIds;
}